requirements.txt
beautifulsoup4
google-api-python-client
google-auth
google-auth-oauthlib
pdfplumber
playwright
requests
